+++ packages/block-editor/src/components/inserter-button/sparkles.js
@@ -1,7 +1,7 @@
 /**
  * WordPress dependencies
  */
-import { SVG, Path } from '@wordpress/primitives';
+import { SVG, Path } from '@wordpress/components';
 
 const sparkles = (
 	<SVG viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">